An In-Memory Mock OVSDB Vtep
This patch contains an in-memory implementation
of a mock vtep for use in testing.

The mock Vtep provides a compliant OvsdbClient
handle that can be used to interact with the
in-memory vtep using the standard ovsdb
interface. Note that only those features required
for the Midonet VTEP support are implemented
in the backend.

The backend implementation provides methods to
examine and manipulate the contents of the
vtep tables.

Ref: MN-2543

Change-Id: I08052f640994b55dd0b3a4ee9271593105d6519c
Signed-off-by: Ernest Artiaga <ernest@midokura.com>